refactor(travel): consolidate neutral flight models into FlightModels.swift (F0.2)
Move the shared flight model vocabulary out of individual component files into
one neutral `Organisms/FlightModels.swift`, per ADR-F7's "model = vocabulary,
configuration = slice" convention. Pure relocation inside the ThemeKit target —
no behavior change, public API byte-identical (API-breakage gate stays green),
and F4's 2.0 extraction becomes a clean `git mv`.

Moved (with full bodies, conformances, nested types, factories):
- FlightStatus (+ label/semantic/icon) ← FlightStatusBadge.swift
- FlightLeg ← FlightCard.swift
- FlightFare ← FlightListItem.swift
- FareLine (+ nested Kind + item/discount/total factories) ← FareSummary.swift
- FareFeatureStatus + FareFeature ← FareFeatureRow.swift

View/style/modifier extensions stayed with their components. FlightModels.swift
needs only `import Foundation` — SemanticColor resolves module-wide via the
`@_exported import ThemeKitCore` in CoreExports.swift.

swift build: clean.
