+++ Sources/ThemeKit/Components/Atoms/FlightStatusBadge.swift
@@ -13,31 +13,6 @@
 
 import SwiftUI
 
-public enum FlightStatus: String, Sendable, CaseIterable {
-    case onTime, boarding, delayed, gateClosed, departed, arrived, cancelled
-
-    var label: String {
-        switch self {
-        case .onTime: "On time"; case .boarding: "Boarding"; case .delayed: "Delayed"
-        case .gateClosed: "Gate closed"; case .departed: "Departed"; case .arrived: "Arrived"; case .cancelled: "Cancelled"
-        }
-    }
-    var semantic: SemanticColor {
-        switch self {
-        case .onTime, .arrived: .success
-        case .boarding, .departed: .info
-        case .delayed, .gateClosed: .warning
-        case .cancelled: .error
-        }
-    }
-    var icon: String {
-        switch self {
-        case .onTime: "checkmark.circle.fill"; case .boarding: "figure.walk"; case .delayed: "clock.fill"
-        case .gateClosed: "lock.fill"; case .departed: "airplane.departure"; case .arrived: "airplane.arrival"; case .cancelled: "xmark.circle.fill"
-        }
-    }
-}
-
 public struct FlightStatusBadge: View {
     @Environment(\.theme) private var theme
